Add files via upload
<!-- 
Hello thank you for the opportunity to work on this project. I have done my best to follow the instructions and I hope i have met the majority of them

task 1: I have fixed the problem with the content not rendering the init() was commeted out.

task 2: setting the tabs to active I have set the active class but I have not worked with tabs an awful lot especially in vanillia js. I am use to setting them more with frameworks such as vue and react

This is somthing  I need to work on in the future and I will be diving deeper and refreshing my knowledge on tabs in vanillia js.

task 3. I feel I have set the loading spinner correctly it does show the spinner when you are selecting the location.

task 4. I have added the api requests for the precipitation data and the temperature data.

task 5. This was one that I got stuck on. I feel I was close with this task but not yet successful. I have never set or added HTML elements in this way before. As mostly I have not needed too.
I did follow how the table was set and tried my best to work it through that way but alas I was unsuccessful. I will be looking into this more in the future.



Overall: I feel this project was really beinfifical. Although I was only able to fully complete 3/5 tasks. I feel I have learnt a great deal. But also I know what I need to work on in the future. I will be looking into tabs in vanillia js and adding HTML elements in this way. I will be looking into this more in the future.
I want to thank you for taking the time to read this and hopefully I have done enough to get either feedback or get a final chat with you all.

I want to thank you for the opportunity and I hope to hear from you soon.

thank you

Daniel

-->